requirements.txt
pyhumps==1.6.1
uvicorn==0.11.7
fastapi==0.60.1
python-multipart==0.0.5